requirements.txt
fastapi
uvicorn
PyJWT
passlib[bcrypt]
python-multipart
slowapi
